Move util functions to shared package
