+++ packages/shared/src/utils/addresses.ts
@@ -16,15 +16,15 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-export * from '@/types'
-export * from '@/api'
-export * from '@/utils'
-export * from '@/errors'
-export * from '@/numbers'
-export * from '@/constants'
-export * from '@/transactions'
-export * from '@/store'
-export * from '@/network'
-export * from '@/currencies'
-export * from '@/assets'
-export * from '@/utils'
+export const findNextAvailableAddressIndex = (startIndex: number, skipIndexes: number[] = []) => {
+  let nextAvailableAddressIndex = startIndex
+
+  do {
+    nextAvailableAddressIndex++
+  } while (skipIndexes.includes(nextAvailableAddressIndex))
+
+  return nextAvailableAddressIndex
+}
+
+export const isAddressIndexValid = (addressIndex: number) =>
+  addressIndex >= 0 && Number.isInteger(addressIndex) && !addressIndex.toString().includes('e')
